feat(map): update the map based on the timestep provided
move each tile back timeStep * 0.06666 in game coordinates

+++ include/Vector2D.hpp
@@ -4,9 +4,9 @@
 
 struct Vector2D
 {
-    uint32_t x, y;
+    double x, y;
 
-    Vector2D(uint32_t pX, uint32_t pY) : x(pX), y(pY)
+    Vector2D(double pX, double pY) : x(pX), y(pY)
     {
     }
 };
